ENH: Create abstract blotter class
