+++ tests/finance/test_blotter.py
@@ -17,7 +17,7 @@ from nose_parameterized import parameterized
 import pandas as pd
 
 from zipline.assets import Equity
-from zipline.finance.blotter import Blotter
+from zipline.finance.simulatedblotter import SimulatedBlotter
 from zipline.finance.cancel_policy import EODCancel, NeverCancel
 from zipline.finance.commission import PerTrade
 from zipline.finance.execution import (
@@ -109,7 +109,7 @@ class BlotterTestCase(WithCreateBarData,
                            (StopLimitOrder(10, 20), 10, 20)])
     def test_blotter_order_types(self, style_obj, expected_lmt, expected_stp):
 
-        blotter = Blotter('daily')
+        blotter = SimulatedBlotter('daily')
 
         blotter.order(self.asset_24, 100, style_obj)
         result = blotter.open_orders[self.asset_24][0]
@@ -118,7 +118,7 @@ class BlotterTestCase(WithCreateBarData,
         self.assertEqual(result.stop, expected_stp)
 
     def test_cancel(self):
-        blotter = Blotter('daily')
+        blotter = SimulatedBlotter('daily')
 
         oid_1 = blotter.order(self.asset_24, 100, MarketOrder())
         oid_2 = blotter.order(self.asset_24, 200, MarketOrder())
@@ -152,7 +152,7 @@ class BlotterTestCase(WithCreateBarData,
         self.assertEqual(list(blotter.open_orders), [self.asset_25])
 
     def test_blotter_eod_cancellation(self):
-        blotter = Blotter('minute', cancel_policy=EODCancel())
+        blotter = SimulatedBlotter('minute', cancel_policy=EODCancel())
 
         # Make two orders for the same asset, so we can test that we are not
         # mutating the orders list as we are cancelling orders
@@ -175,7 +175,7 @@ class BlotterTestCase(WithCreateBarData,
             self.assertEqual(order.status, ORDER_STATUS.CANCELLED)
 
     def test_blotter_never_cancel(self):
-        blotter = Blotter('minute', cancel_policy=NeverCancel())
+        blotter = SimulatedBlotter('minute', cancel_policy=NeverCancel())
 
         blotter.order(self.asset_24, 100, MarketOrder())
 
@@ -189,7 +189,7 @@ class BlotterTestCase(WithCreateBarData,
         self.assertEqual(blotter.new_orders[0].status, ORDER_STATUS.OPEN)
 
     def test_order_rejection(self):
-        blotter = Blotter(self.sim_params.data_frequency)
+        blotter = SimulatedBlotter(self.sim_params.data_frequency)
 
         # Reject a nonexistent order -> no order appears in new_order,
         # no exceptions raised out
@@ -218,7 +218,7 @@ class BlotterTestCase(WithCreateBarData,
 
         # Do it again, but reject it at a later time (after tradesimulation
         # pulls it from new_orders)
-        blotter = Blotter(self.sim_params.data_frequency)
+        blotter = SimulatedBlotter(self.sim_params.data_frequency)
         new_open_id = blotter.order(self.asset_24, 10, MarketOrder())
         new_open_order = blotter.open_orders[self.asset_24][0]
         self.assertEqual(new_open_id, new_open_order.id)
@@ -234,7 +234,7 @@ class BlotterTestCase(WithCreateBarData,
 
         # You can't reject a filled order.
         # Reset for paranoia
-        blotter = Blotter(self.sim_params.data_frequency)
+        blotter = SimulatedBlotter(self.sim_params.data_frequency)
         blotter.slippage_models[Equity] = FixedSlippage()
         filled_id = blotter.order(self.asset_24, 100, MarketOrder())
         filled_order = None
@@ -262,8 +262,8 @@ class BlotterTestCase(WithCreateBarData,
         status indication. When a fill happens, the order should switch
         status to OPEN/FILLED as necessary
         """
-        blotter = Blotter(self.sim_params.data_frequency,
-                          equity_slippage=VolumeShareSlippage())
+        blotter = SimulatedBlotter(self.sim_params.data_frequency,
+                                   equity_slippage=VolumeShareSlippage())
         # Nothing happens on held of a non-existent order
         blotter.hold(56)
         self.assertEqual(blotter.new_orders, [])
@@ -299,8 +299,8 @@ class BlotterTestCase(WithCreateBarData,
             expected_status = ORDER_STATUS.OPEN if expected_open else \
                 ORDER_STATUS.FILLED
 
-            blotter = Blotter(self.sim_params.data_frequency,
-                              equity_slippage=VolumeShareSlippage())
+            blotter = SimulatedBlotter(self.sim_params.data_frequency,
+                                       equity_slippage=VolumeShareSlippage())
             open_id = blotter.order(self.asset_24, order_size, MarketOrder())
             open_order = blotter.open_orders[self.asset_24][0]
             self.assertEqual(open_id, open_order.id)
@@ -322,7 +322,7 @@ class BlotterTestCase(WithCreateBarData,
             self.assertEqual(filled_order.open_amount, expected_open)
 
     def test_prune_orders(self):
-        blotter = Blotter(self.sim_params.data_frequency)
+        blotter = SimulatedBlotter(self.sim_params.data_frequency)
 
         blotter.order(self.asset_24, 100, MarketOrder())
         open_order = blotter.open_orders[self.asset_24][0]
@@ -349,8 +349,8 @@ class BlotterTestCase(WithCreateBarData,
         Ensure the effect of order_batch is the same as multiple calls to
         order.
         """
-        blotter1 = Blotter(self.sim_params.data_frequency)
-        blotter2 = Blotter(self.sim_params.data_frequency)
+        blotter1 = SimulatedBlotter(self.sim_params.data_frequency)
+        blotter2 = SimulatedBlotter(self.sim_params.data_frequency)
         for i in range(1, 4):
             order_arg_lists = [
                 (self.asset_24, i * 100, MarketOrder()),
@@ -377,7 +377,7 @@ class BlotterTestCase(WithCreateBarData,
                                  blotter2.open_orders[asset][i-1].id)
 
     def test_slippage_and_commission_dispatching(self):
-        blotter = Blotter(
+        blotter = SimulatedBlotter(
             self.sim_params.data_frequency,
             equity_slippage=FixedSlippage(spread=0.0),
             future_slippage=FixedSlippage(spread=2.0),
