+++ zipline/algorithm.py
@@ -63,6 +63,7 @@ from zipline.errors import (
     UnsupportedOrderParameters,
 )
 from zipline.finance.trading import TradingEnvironment
+from zipline.finance.simulatedblotter import SimulatedBlotter
 from zipline.finance.blotter import Blotter
 from zipline.finance.controls import (
     LongOnly,
@@ -326,11 +327,18 @@ class TradingAlgorithm(object):
         self.blotter = kwargs.pop('blotter', None)
         self.cancel_policy = kwargs.pop('cancel_policy', NeverCancel())
         if not self.blotter:
-            self.blotter = Blotter(
+            self.blotter = SimulatedBlotter(
                 data_frequency=self.data_frequency,
                 # Default to NeverCancel in zipline
                 cancel_policy=self.cancel_policy,
             )
+
+        if not issubclass(self.blotter.__class__, Blotter):
+            raise TypeError("Blotter specified is of type %s "
+                            "which is not a subclass of "
+                            "zipline.finance.blotter.Blotter" %
+                            self.blotter.__module__ + "." +
+                            self.blotter.__class__.__name__)
 
         # The symbol lookup date specifies the date to use when resolving
         # symbols to sids, and can be set using set_symbol_lookup_date()
